requirements.txt
pillow
pylablib
tqdm